requirements.txt
Flask==1.1.2
Flask-Script==2.0.6
WTForms==2.3.3
Flask-PyMongo==2.3.0
python-decouple>=3.1,<4